# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: WIAL Backend API
  version: 1.0.0
  description: |
    OpenAPI documentation for the WIAL backend.

    Public auth is login-only. User creation is restricted to authenticated
    administrators through the managed users endpoint.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/users:
    get:
      tags: [Users]
      summary: List managed users
      description: Chapter leads are scoped to their assigned chapter. Super-admins can optionally filter by `chapter_id`.
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: chapter_id
          schema:
            type: string
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Users]
      summary: Create a managed user
      description: |
        Creates a `chapter_lead`, `coach`, or `content_creator`.
        `super_admin` users are bootstrapped from env only and cannot be created through the API.
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request, invalid role, or missing chapter assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: email already exists
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/users/{id}:
    get:
      tags: [Users]
      summary: Get a managed user
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Users]
      summary: Patch a managed user
      operationId: patchUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPatchRequest'
      responses:
        '200':
          description: User patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request, invalid role, or no fields to update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: email already exists
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Users]
      summary: Delete a managed user
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: User deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/testimonials:
    get:
      tags: [Testimonials]
      summary: List testimonials
      operationId: listTestimonials
      parameters:
        - in: query
          name: chapter_id
          schema:
            type: string
        - in: query
          name: program_id
          schema:
            type: string
      responses:
        '200':
          description: Testimonial list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestimonialListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Testimonials]
      summary: Create a testimonial
      operationId: createTestimonial
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestimonialCreateRequest'
      responses:
        '201':
          description: Testimonial created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Testimonial'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/testimonials/{id}:
    get:
      tags: [Testimonials]
      summary: Get a testimonial
      operationId: getTestimonial
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Testimonial found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Testimonial'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Testimonials]
      summary: Patch a testimonial
      operationId: patchTestimonial
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestimonialPatchRequest'
      responses:
        '200':
          description: Testimonial patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Testimonial'
        '400':
          description: Invalid request or no fields to update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Testimonials]
      summary: Delete a testimonial
      operationId: deleteTestimonial
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Testimonial deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed or missing bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated user lacks the required role or assignment scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        id: { type: string }
        email:
          type: string
          format: email
        role:
          type: string
          enum: [super_admin, chapter_lead, coach, content_creator]
        chapter_id:
          type: string
          nullable: true
        chapter_name:
          type: string
          nullable: true
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, email, role, created_at, updated_at]
    UserCreateRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [chapter_lead, coach, content_creator]
        chapter_id:
          type: string
      required: [email, password, role, chapter_id]
    UserPatchRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [super_admin, chapter_lead, coach, content_creator]
        chapter_id:
          type: string
    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
      required: [data]
...
    Testimonial:
      type: object
      properties:
        id: { type: string }
        author_name: { type: string }
        author_title: { type: string }
        author_company: { type: string, nullable: true }
        author_image_url: { type: string, nullable: true }
        content: { type: string }
        rating: { type: integer, nullable: true }
        chapter_id: { type: string, nullable: true }
        program_id: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, author_name, author_title, content, created_at, updated_at]
    TestimonialCreateRequest:
      type: object
      properties:
        author_name: { type: string }
        author_title: { type: string }
        author_company: { type: string, nullable: true }
        author_image_url: { type: string, nullable: true }
        content: { type: string }
        rating: { type: integer, nullable: true }
        chapter_id: { type: string, nullable: true }
        program_id: { type: string, nullable: true }
      required: [author_name, author_title, content]
    TestimonialPatchRequest:
      type: object
      properties:
        author_name: { type: string }
        author_title: { type: string }
        author_company: { type: string }
        author_image_url: { type: string }
        content: { type: string }
        rating: { type: integer }
        chapter_id: { type: string }
        program_id: { type: string }
    TestimonialListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Testimonial'
      required: [data]
...