# backend/internal/models/user.go
package models

import "time"

const (
	RoleSuperAdmin     = "super_admin"
	RoleChapterLead    = "chapter_lead"
	RoleCoach          = "coach"
	RoleContentCreator = "content_creator"
)

type User struct {
	ID           string    `json:"id"`
	Email        string    `json:"email"`
	PasswordHash string    `json:"-"`
	Role         string    `json:"role"`
	ChapterID    *string   `json:"chapter_id,omitempty"`
	ChapterName  *string   `json:"chapter_name,omitempty"`
	CreatedAt    time.Time `json:"created_at"`
	UpdatedAt    time.Time `json:"updated_at"`
}

type UserRegisterRequest struct {
	Email     string  `json:"email" binding:"required,email"`
	Password  string  `json:"password" binding:"required,min=8"`
	Role      string  `json:"role" binding:"required"`
	ChapterID *string `json:"chapter_id"`
}

type UserCreateRequest struct {
	Email     string  `json:"email" binding:"required,email"`
	Password  string  `json:"password" binding:"required,min=8"`
	Role      string  `json:"role" binding:"required"`
	ChapterID *string `json:"chapter_id"`
}

type UserPatchRequest struct {
	Email     *string `json:"email"`
	Password  *string `json:"password"`
	Role      *string `json:"role"`
	ChapterID *string `json:"chapter_id"`
}

type UserListResponse struct {
	Data []User `json:"data"`
}

type UserLoginRequest struct {
	Email    string `json:"email" binding:"required,email"`
	Password string `json:"password" binding:"required"`
}

type AuthResponse struct {
	Token string `json:"token"`
	User  User   `json:"user"`
}